@@ -1,32 +1,27 @@
 [package]
 name = "memoires-api"
 version = "0.1.0"
-edition = "2021"
+edition = "2024"
 
 [workspace]
 members = [".", "migration"]
 
 [dependencies]
-migration = { path = "migration" }
-axum = { version = "0.7.5" }
+async-trait = "0.1.85"
+axum = { version = "0.8.0" }
+axum-keycloak-auth = "0.8.1"
+base64 = "0.22.1"
+byteorder = "1.5.0"
+chrono = { version = "0.4.40", features = ["serde"] }
+crudcrate = "0.2.3"
+# crudcrate = { path = "../crudcrate" }
 dotenvy = "0.15.7"
-serde = { version = "1.0.209", features = ["derive"] }
-serde_json = "1.0.128"
-# sqlx = { version = "0.7.4", features = [
-#     "postgres",
-#     "runtime-tokio-native-tls",
-#     "macros",
-#     "uuid",
-#     "chrono",
-#     # "sqlite",
-# ], default-features = false }
-tokio = { version = "1.40.0", features = ["full"] }
-hyper = "1.4.1"
-tracing = "0.1.40"
-tracing-subscriber = "0.3.18"
-uuid = { version = "1.10.0", features = ["serde", "v4", "fast-rng"] }
-chrono = { version = "0.4.38", features = ["serde"] }
-sea-orm = { version = "1.0.1", features = [
+futures = "0.3.31"
+gpx = "0.10.0"
+hyper = "1.5.2"
+migration = { path = "migration" }
+rand = "0.9.0"
+sea-orm = { version = "1.1.6", features = [
     "sqlx-postgres",
     "runtime-tokio-rustls",
     "macros",
@@ -34,25 +29,31 @@ sea-orm = { version = "1.0.1", features = [
     "with-chrono",
     "with-json",
 ], default-features = false }
-sea-orm-migration = "1.0.1"
-# proj = "0.27.2"
-wkt = { version = "0.11.0", features = ["serde", "geo-types"] }
-geozero = { version = "0.13.0", features = ["with-postgis-sqlx"] }
-geo-types = { version = "0.7.13", features = ["serde"] }
-utoipa = { version = "4.2.3", features = ["axum_extras", "uuid"] }
-sea-query = "0.31.0"
-# utoipa-axum = "0.1.0-beta.2"
-utoipa-swagger-ui = { version = "7.1.0", features = ["axum"] }
-utoipa-redoc = { version = "4.0.0", features = ["axum"] }
-utoipa-rapidoc = { version = "4.0.0", features = ["axum"] }
-utoipa-scalar = { version = "0.1.0", features = ["axum"] }
-wkb = "0.7.1"
-byteorder = "1.5.0"
-lttb = "0.2.0"
-postgis = "0.9.0"
-# geo = "0.28.0"
+sea-orm-migration = "1.1.7"
+serde = { version = "1.0.217", features = ["derive"] }
+serde_json = "1.0.137"
 serde_with = "3.10.0"
+tokio = { version = "1.43.0", features = ["full"] }
+tracing = "0.1.41"
+tracing-subscriber = "0.3.19"
+utoipa = { version = "5.3.1", features = ["axum_extras", "uuid", "chrono"] }
+utoipa-axum = "0.2.0"
+utoipa-scalar = { version = "0.3.0", features = ["axum"] }
+uuid = { version = "1.13.2", features = ["serde", "v4", "fast-rng"] }
 
 [dev-dependencies]
-reqwest = { version="0.12.7", features = ["json"] }
-tokio = { version = "1.40.0", features = ["full"] }
+async-std = { version = "1.13", features = ["attributes"] }
+rstest = "0.25.0"
+sea-orm = { version = "1.0.1", features = [
+    "sqlx-postgres",
+    "sqlx-sqlite",
+    "runtime-tokio-rustls",
+    "macros",
+    "with-uuid",
+    "with-chrono",
+    "with-json",
+], default-features = false }
+tower = { version = "0.5.2", features = ["util"] }
+
+[lints.clippy]
+pedantic = "warn"
